open file upload settings after files selected
